Fix subtraction with overflow bug
If we don't find a newline then that means this variable is on the
first line, in that case we need to take start_position as our
indent_level and not `start_position - 1` (newline_position will be 0
because of unwrap_or).

Taking `start_position - 1` is going to cause subtraction with
overflow if the variable was the first thing in the file and it is
wrong as well.

```
  <!--% var %-->
```

`indent_level` for this should be 2, `start_position` is 2 as well.